+++ Mslm.Tests/EmailVerifyTests.cs
@@ -9,18 +9,18 @@ namespace Mslm.Tests
         [Fact]
         public async Task TestSingleVerify()
         {
-            // Arrange
-            // to run this test, add your MSLM Api Key to environment variable
+            // To run this test, add your Mslm API Key to environment variable
             // named "MSLM_API_KEY", or initialize your key string directly.
             string token = Environment.GetEnvironmentVariable("MSLM_API_KEY") ?? "default-api-key";
 
+            // Set up client.
             var emailVerifyClient = new EmailVerify(token);
             string emailToVerify = "abc@gmail.com";
 
-            // Act
+            // Get result.
             var result = await emailVerifyClient.SingleVerify(emailToVerify);
 
-            // Assert
+            // Assert.
             Assert.NotNull(result);
         }
     }