+++ samples/OtpSendSample/Program.cs
@@ -17,7 +17,7 @@ namespace OtpSendSample
                 ExpireSeconds = 300
             };
 
-            // to use this sample, add your MSLM Api Key to environment variable
+            // To use this sample, add your Mslm API Key to environment variable
             // named "MSLM_API_KEY", or initialize your key string directly.
             string token = Environment.GetEnvironmentVariable("MSLM_API_KEY") ?? "default-api-key";