+++ samples/OtpTokenVerifySample/Program.cs
@@ -16,7 +16,7 @@ namespace OtpTokenVerifySample
                 Consume = true
             };
 
-            // to use this sample, add your MSLM Api Key to environment variable
+            // To use this sample, add your Mslm API Key to environment variable
             // named "MSLM_API_KEY", or initialize your key string directly.
             string token = Environment.GetEnvironmentVariable("MSLM_API_KEY") ?? "default-api-key";